+++ src/components/split-button/split-button-primary-action.component.ts
@@ -1,0 +1,7 @@
+import {Component} from '@angular/core';
+
+@Component({
+    selector: 'gtx-split-button-primary-action',
+    template: '<ng-content></ng-content>'
+})
+export class SplitButtonPrimaryAction { }